Rename shops to ChicagoBikeShops at /chicago-shops And drop from header
