+++ src/router/index.ts
@@ -10,7 +10,7 @@ import SocialImageBike from '../components/SocialImageBike.vue'
 import SocialImagePage from '../components/SocialImagePage.vue'
 import AdSquare from '../components/AdSquare.vue'
 import AdsPage from '../components/AdsPage.vue'
-import ShopsPage from '../components/ShopsPage.vue'
+import ChicagoBikeShopsPage from '../components/ChicagoBikeShopsPage.vue'
 import AllBikesPage from '../components/AllBikesPage.vue'
 import AdminPage from '../components/AdminPage.vue'
 import NotFoundPage from '../components/NotFoundPage.vue'
@@ -58,9 +58,9 @@ const routes = [
     meta: { title: PageMeta.storage.title }
   },
   {
-    path: '/shops',
-    name: 'Shops',
-    component: ShopsPage,
+    path: '/chicago-shops',
+    name: 'ChicagoBikeShops',
+    component: ChicagoBikeShopsPage,
     meta: { title: PageMeta.shops.title }
   },
   {
